requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
instagram-private-api==1.6.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.4.2